+++ demo_cli.py
@@ -9,7 +9,7 @@ import torch
 
 from encoder import inference as encoder
 from encoder.params_model import model_embedding_size as speaker_embedding_size
-from synthesizer.models.tacotron_tweaked.inference import Synthesizer
+from synthesizer.models.tacotron.inference import Synthesizer
 from utils.argutils import print_args
 from vocoder import inference as vocoder
 from synthesizer.g2p import g2p_main
