+++ synthesizer/g2p/__init__.py
@@ -170,7 +170,7 @@ def init(dl_logger_=None):
     regexPattern = '|'.join('(?={})'.format(re.escape(delim)) for delim in delims)
     if dl_logger_ is None:
         try:
-            from synthesizer.models.tacotron_tweaked.train import dl_logger
+            from synthesizer.models.tacotron.train import dl_logger
         except Exception as e:
             print(e)
             dl_logger = ShortLogger()
