+++ synthesizer/models/tacotron/inference.py
@@ -67,9 +67,9 @@ class Synthesizer:
 
         if self.verbose:
             print("Loaded synthesizer \"%s\" trained to step %d" % (
-            self.model_fpath.name, self._model.state_dict()["step"]))
+                self.model_fpath.name, self._model.state_dict()["step"]))
 
-    def synthesize_spectrograms(self, texts,
+    def synthesize_spectrograms(self, text,
                                 embeddings: Union[np.ndarray, List[np.ndarray]],
                                 return_alignments=False):
         """
@@ -87,46 +87,29 @@ class Synthesizer:
         if not self.is_loaded():
             self.load()
 
-        texts = [item for sublist in texts for item in sublist]
-        inputs = text_to_sequence(texts)
+        inputs = text_to_sequence(text)
 
         if not isinstance(embeddings, list):
             embeddings = [embeddings]
 
         # Batch inputs
-        batched_inputs = [inputs[i:i + hparams.synthesis_batch_size]
-                          for i in range(0, len(inputs), hparams.synthesis_batch_size)]
         batched_embeds = [embeddings[i:i + hparams.synthesis_batch_size]
                           for i in range(0, len(embeddings), hparams.synthesis_batch_size)]
 
         specs = []
-        alignments = None
-        for i, batch in enumerate(batched_inputs, 1):
-            if self.verbose:
-                print(f"\n| Generating {i}/{len(batched_inputs)}")
-
-            # Pad texts so they are all the same length
-            print(batch)
-            text_lens = [len(text) for text in batch]
-            max_text_len = max(text_lens)
-            chars = [pad1d(text, max_text_len) for text in batch]
-            chars = np.stack(chars)
-
-            # Stack speaker embeddings into 2D array for batch processing
-            speaker_embeds = np.stack(batched_embeds[i - 1])
-
-            # Convert to tensor
-            chars = torch.tensor(chars).long().to(self.device)
-            speaker_embeddings = torch.tensor(speaker_embeds).float().to(self.device)
-
-            # Inference
-            _, mels, alignments = self._model.generate(chars, speaker_embeddings)
-            mels = mels.detach().cpu().numpy()
-            for m in mels:
-                # Trim silence from end of each spectrogram
-                while np.max(m[:, -1]) < hparams.tts_stop_threshold:
-                    m = m[:, :-1]
-                specs.append(m)
+        batch = torch.tensor(inputs).unsqueeze(0).to(self.device)
+        # print(batch)
+        speaker_embeds = np.stack(batched_embeds[0])
+        speaker_embeddings = torch.tensor(speaker_embeds).float().to(self.device)
+        # print(batch.shape)
+        # Inference
+        _, mels, alignments = self._model.generate(batch, speaker_embeddings)
+        mels = mels.detach().cpu().numpy()
+        for m in mels:
+            # Trim silence from end of each spectrogram
+            while np.max(m[:, -1]) < hparams.tts_stop_threshold:
+                m = m[:, :-1]
+            specs.append(m)
 
         if self.verbose:
             print("\n\nDone.\n")
@@ -168,3 +151,7 @@ class Synthesizer:
 
 def pad1d(x, max_len, pad_value=0):
     return np.pad(x, (0, max_len - len(x)), mode="constant", constant_values=pad_value)
+
+
+def pad2d(x, max_len, pad_value=0):
+    return np.pad(x, ((0, 0), (0, max_len - x.shape[-1])), mode="constant", constant_values=pad_value)
