+++ synthesizer/models/tacotron/synthesizer_dataset.py
@@ -1,18 +1,19 @@
+from pathlib import Path
+
+import numpy as np
 import torch
 from torch.utils.data import Dataset
-import numpy as np
-from pathlib import Path
+
 from synthesizer.utils.text import text_to_sequence
-from synthesizer.g2p import g2p_main
 
 
 class SynthesizerDataset(Dataset):
     def __init__(self, metadata_fpath: Path, mel_dir: Path, embed_dir: Path, hparams):
         print("Using inputs from:\n\t%s\n\t%s\n\t%s" % (metadata_fpath, mel_dir, embed_dir))
-        
+
         with metadata_fpath.open("r", encoding="utf8") as metadata_file:
             metadata = [line.split("|") for line in metadata_file]
-        
+
         mel_fnames = [x[1] for x in metadata if int(x[4])]
         mel_fpaths = [mel_dir.joinpath(fname) for fname in mel_fnames]
         embed_fnames = [x[2] for x in metadata if int(x[4])]
@@ -22,28 +23,39 @@ class SynthesizerDataset(Dataset):
         self.metadata = metadata
         self.hparams = hparams
         self.debug = False  # hardcoded but if you can't code you don't need debug..right?
-        
+
         print("Found %d samples" % len(self.samples_fpaths))
-    
-    def __getitem__(self, index):  
-        # Sometimes index may be a list of 2 (not sure why this happens)
-        # If that is the case, return a single item corresponding to first element in index
-        if index is list:
-            index = index[0]
-            print("what the fuck honestly")
+
+    def __getitem__(self, index):
         mel_path, embed_path = self.samples_fpaths[index]
-        mel = np.load(mel_path).T.astype(np.float32)
+        try:
+
+            mel = np.load(mel_path, allow_pickle=True).T.astype(np.float32)
+        except:
+            try:
+                mel = np.load(str(mel_path), allow_pickle=True).T.astype(np.float32)
+            except:
+                mel_path, embed_path = self.samples_fpaths[index+1]
+                mel = np.load(mel_path, allow_pickle=True).T.astype(np.float32)
         # Load the embed
         embed = np.load(embed_path)
-        # Get the text and (not) clean it
-        text = text_to_sequence(g2p_main(self.samples_texts[index].lower()))[:-1]
+
+        text = text_to_sequence(self.samples_texts[index].lower())
         # Convert the list returned by text_to_sequence to a numpy array
         text = np.asarray(text).astype(np.int32)
 
-        return text, mel.astype(np.float32), embed.astype(np.float32), index
+        return text, mel, embed.astype(np.float32), index
 
     def __len__(self):
         return len(self.samples_fpaths)
+
+
+def pad1d(x, max_len, pad_value=0):
+    return np.pad(x, (0, max_len - len(x)), mode="constant", constant_values=pad_value)
+
+
+def pad2d(x, max_len, pad_value=0):
+    return np.pad(x, ((0, 0), (0, max_len - x.shape[-1])), mode="constant", constant_values=pad_value)
 
 
 def collate_synthesizer(batch, r, hparams):
@@ -56,11 +68,11 @@ def collate_synthesizer(batch, r, hparams):
 
     # Mel spectrogram
     spec_lens = [x[1].shape[-1] for x in batch]
-    max_spec_len = max(spec_lens) + 1 
+    max_spec_len = max(spec_lens) + 1
     if max_spec_len % r != 0:
-        max_spec_len += r - max_spec_len % r 
+        max_spec_len += r - max_spec_len % r
 
-    # WaveRNN mel spectrograms are normalized to [0, 1] so zero padding adds silence
+        # WaveRNN mel spectrograms are normalized to [0, 1] so zero padding adds silence
     # By default, SV2TTS uses symmetric mels, where -1*max_abs_value is silence.
     if hparams.symmetric_mels:
         mel_pad_value = -1 * hparams.max_abs_value
@@ -76,7 +88,6 @@ def collate_synthesizer(batch, r, hparams):
     # Index (for vocoder preprocessing)
     indices = [x[3] for x in batch]
 
-
     # Convert all to tensor
     chars = torch.tensor(chars).long()
     mel = torch.tensor(mel)
@@ -84,8 +95,6 @@ def collate_synthesizer(batch, r, hparams):
 
     return chars, mel, embeds, indices
 
-def pad1d(x, max_len, pad_value=0):
-    return np.pad(x, (0, max_len - len(x)), mode="constant", constant_values=pad_value)
 
-def pad2d(x, max_len, pad_value=0):
-    return np.pad(x, ((0, 0), (0, max_len - x.shape[-1])), mode="constant", constant_values=pad_value)
+def flat(xss):
+    return [x for xs in xss for x in xs]
