--- synthesizer/models/tacotron_tweaked/__init__.py
@@ -1,1 +1,0 @@
-#
